+++ index.js
@@ -1,5 +1,5 @@
-import { DateTime } from './node_modules/luxon/src/luxon.js';
 import { showBookList, addBook } from './modules/book.js';
+import generateDate from './modules/date.js';
 
 const bookForm = document.querySelector('#bookForm');
 const bookList = document.querySelector('#bookList');
@@ -61,20 +61,4 @@ const show = (element, link) => {
 });
 
 // add date
-let suffix;
-const date = DateTime.now().toFormat('d');
-
-if (+date[1] === 1 && +date !== 11) {
-  suffix = 'st';
-} else if (+date[1] === 2 && +date !== 12) {
-  suffix = 'nd';
-} else if (+date[1] === 3 && +date !== 13) {
-  suffix = 'rd';
-} else {
-  suffix = 'th';
-}
-
-setInterval(() => {
-  const now = DateTime.now().toFormat(`MMMM dd'${suffix}' yyyy', ' hh:mm:ss a`);
-  document.getElementById('date').innerHTML = now;
-}, 1000);
+generateDate();
